requirements.txt
fastapi
uvicorn
jinja2
openai
faiss-cpu
requests
beautifulsoup4
lxml
pyyaml
geoip2
numpy
pytest
